refactor: move to the helpers folder the email-parser
